refactor and create relational test server
