+++ legend-engine-server/src/test/java/org/finos/legend/engine/server/test/pureClient/other/Test_Engine_UsingPureClient.java
@@ -17,6 +17,7 @@ package org.finos.legend.engine.server.test.pureClient.other;
 import junit.framework.Test;
 import junit.framework.TestSuite;
 import org.finos.legend.engine.server.test.shared.PureTestHelper;
+import org.finos.legend.engine.server.test.shared.PureWithEngineHelper;
 import org.finos.legend.pure.m3.execution.test.TestCollection;
 import org.finos.legend.pure.runtime.java.compiled.execution.CompiledExecutionSupport;
 
@@ -27,7 +28,7 @@ public class Test_Engine_UsingPureClient extends TestSuite
     public static Test suite()
     {
         return wrapSuite(
-                () -> PureTestHelper.initClientVersionIfNotAlreadySet("vX_X_X"),
+                () -> PureWithEngineHelper.initClientVersionIfNotAlreadySet("vX_X_X"),
                 () ->
                 {
                     CompiledExecutionSupport executionSupport = PureTestHelper.getClassLoaderExecutionSupport();
