--- legend-engine-server/src/test/java/org/finos/legend/engine/server/test/shared/TestServerConfiguration.java
@@ -1,31 +1,0 @@
-//  Copyright 2022 Goldman Sachs
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-
-package org.finos.legend.engine.server.test.shared;
-
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.DatabaseType;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.store.relational.connection.RelationalDatabaseConnection;
-import org.finos.legend.engine.server.ServerConfiguration;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public class TestServerConfiguration extends ServerConfiguration
-{
-    public List<DatabaseType> testConnectionsToEnable = new ArrayList<>();
-    public Map<DatabaseType, RelationalDatabaseConnection> staticTestConnections = new HashMap<>();
-    public Map<DatabaseType, String> dynamicTestConnectionCreators = new HashMap<>();
-}
